requirements.txt
mutagen
imageio[avif]
Pillow